feat: 공지사항 페이지 연결 및 일부 디자인 수정
